Support for dependencies of code examples
A code example can now declare on which other (max 1) code example it
depends. To make such a dependency, the latter code example must be
given a name, and the former should have a name (because of parsing
rules) and a dependency declaration. Names are still optional, and only
required for dependencies.

Syntax:

```JavaScript[example1]
...

and

```JavaScript[example2][example1]
...

The second code example will depend on the first one.

When evaluating (in the browser) a code example with a dependency, the
initial contents of the dependency will be evaluated prior to evaluating
the current code example. Output (including `println`) in the dependency
will be suppressed.